requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile requirements.in
#
beautifulsoup4==4.10.0
    # via -r requirements.in
black==21.12b0
    # via -r requirements.in
cachetools==4.2.4
    # via google-auth
certifi==2021.10.8
    # via requests
charset-normalizer==2.0.10
    # via requests
click==8.0.3
    # via
    #   black
    #   pip-tools
    #   typer
colorama==0.4.4
    # via typer
flake8==4.0.1
    # via -r requirements.in
google-api-core==2.4.0
    # via google-api-python-client
google-api-python-client==2.36.0
    # via -r requirements.in
google-auth==2.3.3
    # via
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
    #   google-auth-oauthlib
google-auth-httplib2==0.1.0
    # via
    #   -r requirements.in
    #   google-api-python-client
google-auth-oauthlib==0.4.6
    # via -r requirements.in
googleapis-common-protos==1.54.0
    # via google-api-core
httplib2==0.20.2
    # via
    #   google-api-python-client
    #   google-auth-httplib2
idna==3.3
    # via requests
jinja2==3.0.3
    # via -r requirements.in
markupsafe==2.0.1
    # via jinja2
mccabe==0.6.1
    # via flake8
mypy-extensions==0.4.3
    # via black
oauthlib==3.1.1
    # via requests-oauthlib
pathspec==0.9.0
    # via black
pep517==0.12.0
    # via pip-tools
pip-tools==6.4.0
    # via -r requirements.in
platformdirs==2.4.1
    # via black
protobuf==3.19.3
    # via
    #   google-api-core
    #   googleapis-common-protos
pyasn1==0.4.8
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.2.8
    # via google-auth
pycodestyle==2.8.0
    # via flake8
pyflakes==2.4.0
    # via flake8
pyparsing==3.0.7
    # via httplib2
pyperclip==1.8.2
    # via -r requirements.in
python-dateutil==2.8.2
    # via -r requirements.in
python-dotenv==0.19.2
    # via -r requirements.in
python-frontmatter==1.0.0
    # via -r requirements.in
pyyaml==6.0
    # via python-frontmatter
requests==2.27.1
    # via
    #   -r requirements.in
    #   google-api-core
    #   requests-oauthlib
requests-oauthlib==1.3.0
    # via google-auth-oauthlib
rsa==4.8
    # via google-auth
shellingham==1.4.0
    # via typer
six==1.16.0
    # via
    #   google-auth
    #   google-auth-httplib2
    #   python-dateutil
soupsieve==2.3.1
    # via beautifulsoup4
tomli==1.2.3
    # via
    #   black
    #   pep517
typer[all]==0.4.0
    # via -r requirements.in
typing-extensions==4.0.1
    # via black
uritemplate==4.1.1
    # via google-api-python-client
urllib3==1.26.8
    # via requests
wheel==0.37.1
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
